requirements.txt
# =============================================================================
# AeroGuardian - Python Dependencies
# =============================================================================
# Python 3.10+ Required
# Author: AeroGuardian Member
# Date: 2026-01-30
# Version: 1.0 (Industry-Standard Refactor)
#
# Setup:
#   python -m venv venv
#   venv\Scripts\activate
#   pip install -r requirements.txt
# =============================================================================

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# Scientific Computing & Data
numpy>=1.26.0,<2.0.0              # Array operations
pandas>=2.3.0                      # DataFrame operations

# Environment
python-dotenv>=1.2.0               # Load .env variables

# =============================================================================
# LLM INTEGRATION
# =============================================================================

# OpenAI & DSPy
openai>=2.14.0                     # OpenAI API client
dspy-ai>=3.1.0                     # DSPy framework for structured LLM output
litellm>=1.80.0                    # LLM provider abstraction

# Async HTTP (required by litellm)
aiohttp>=3.10.0                    # Async HTTP client

# =============================================================================
# REPORTING & VISUALIZATION
# =============================================================================

# Excel Reports
openpyxl>=3.1.0                    # Excel file generation

# PDF Reports
reportlab>=4.4.0                   # Professional PDF generation

# Data Validation
pydantic>=2.0.0                    # Data validation

# =============================================================================
# PX4 SITL INTEGRATION
# =============================================================================

# MAVLink Communication
mavsdk>=3.10.0                     # Async PX4 mission control
pymavlink>=2.4.0                   # MAVLink protocol

# HTTP Requests (for geocoding)
requests>=2.32.0                   # HTTP client

# =============================================================================
# VALIDATION & EVALUATION
# =============================================================================

# Semantic Similarity
sentence-transformers>=5.0.0       # Text embeddings for validation

# =============================================================================
# OPTIONAL: SCIENTIFIC ANALYSIS
# =============================================================================

# Machine Learning (for evaluation metrics)
scikit-learn>=1.3.0                # ML metrics, preprocessing